Changed - Moved BatchOperations into Messages folder - Move request and response models into Models folder - Renamed Exception folder from singular to plural
