+++ dotnet/Xrm.Crm.WebApi/Interfaces/IWebApiBatchOperations.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using Xrm.Crm.WebApi.BatchOperations;
+using Xrm.Crm.WebApi.Messages.BatchOperations;
 using Xrm.Crm.WebApi.Models;
 using Xrm.Crm.WebApi.Response;
