+++ dotnet/Xrm.Crm.WebApi/Messages/Actions/CloseQuoteRequest.cs
@@ -1,5 +1,6 @@
 ﻿using Xrm.Crm.WebApi.Interfaces;
 using Xrm.Crm.WebApi.Models;
+using Xrm.Crm.WebApi.Models.Requests;
 
 namespace Xrm.Crm.WebApi.Messages.Actions
 {
