+++ dotnet/Xrm.Crm.WebApi/Messages/Actions/MergeRequest.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using Newtonsoft.Json.Linq;
-using Xrm.Crm.WebApi.Exception;
+using Xrm.Crm.WebApi.Exceptions;
 using Xrm.Crm.WebApi.Models;
 using Xrm.Crm.WebApi.Request;
