+++ dotnet/Xrm.Crm.WebApi/WebApiBatchOperations.cs
@@ -3,9 +3,10 @@ using System.Collections.Generic;
 using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
-using Xrm.Crm.WebApi.BatchOperations;
 using Xrm.Crm.WebApi.Interfaces;
+using Xrm.Crm.WebApi.Messages.BatchOperations;
 using Xrm.Crm.WebApi.Models;
+using Xrm.Crm.WebApi.Request;
 using Xrm.Crm.WebApi.Response;
 
 namespace Xrm.Crm.WebApi
